agents/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Google ADK & AI
google-adk[database]>=1.16.0
google-generativeai>=0.8.0

# Database
supabase>=2.0.0